Add a template file so we can separate our logic from the generated files.
Without a template, any rebuild of the API will throw away our changes which
is a tad annoying. This way we can keep the logic in a separate python module
that is called from the generated code (and has basically a 1:1 mapping for
function names).

Solution from
https://stackoverflow.com/questions/45680298/cleanest-way-to-glue-generated-flask-app-code-swagger-codegen-to-backend-imple/47554626#47554626

the controller.mustache directly downloaded from github.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>